+++ docs/about/whats_new.rst
@@ -26,6 +26,7 @@ v1.9.next
 - Virtual product tests using odc-geo GridSpec (:pull:`1512`)
 - New Configuration API, as per ODC-EP10 (:pull:`1505`)
 - Alembic migrations for postgis driver (:pull:`1520`)
+- EP08 lineage extensions/changes to datasets.get(). (:pull:`1530`)
 
 
 v1.8.next